Добавлены контроллеры, ДТО, мэпперы для сущностей отделение, банкомат, аудит

+++ public-info/src/main/java/com/bank/publicinfo/repository/AtmRepository.java
@@ -3,5 +3,7 @@ package com.bank.publicinfo.repository;
 import com.bank.publicinfo.entity.Atm;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AtmRepository extends JpaRepository<Atm, Long> {
+import java.math.BigInteger;
+
+public interface AtmRepository extends JpaRepository<Atm, BigInteger> {
 }
