+++ public-info/src/main/java/com/bank/publicinfo/repository/AuditRepository.java
@@ -3,5 +3,7 @@ package com.bank.publicinfo.repository;
 import com.bank.publicinfo.entity.Audit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditRepository extends JpaRepository<Audit, Long> {
+import java.math.BigInteger;
+
+public interface AuditRepository extends JpaRepository<Audit, BigInteger> {
 }
