+++ public-info/src/main/java/com/bank/publicinfo/repository/BranchRepository.java
@@ -3,5 +3,7 @@ package com.bank.publicinfo.repository;
 import com.bank.publicinfo.entity.Branch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BranchRepository extends JpaRepository<Branch, Long> {
+import java.math.BigInteger;
+
+public interface BranchRepository extends JpaRepository<Branch, BigInteger> {
 }
