+++ public-info/src/main/java/com/bank/publicinfo/repository/LicenseRepository.java
@@ -3,5 +3,7 @@ package com.bank.publicinfo.repository;
 import com.bank.publicinfo.entity.License;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LicenseRepository extends JpaRepository<License, Long> {
+import java.math.BigInteger;
+
+public interface LicenseRepository extends JpaRepository<License, BigInteger> {
 }
